Move packages to sdk/resourcemanagerhybrid; restore previous

+++ sdk/resourcemanager/azure-resourcemanager-network/src/main/java/com/azure/resourcemanager/network/fluent/models/PacketCaptureInner.java
@@ -5,33 +5,27 @@
 package com.azure.resourcemanager.network.fluent.models;
 
 import com.azure.core.annotation.Fluent;
-import com.azure.core.annotation.Immutable;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.resourcemanager.network.models.PacketCaptureFilter;
 import com.azure.resourcemanager.network.models.PacketCaptureStorageLocation;
-import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
 
-/**
- * Parameters that define the create packet capture operation.
- */
+/** Parameters that define the create packet capture operation. */
 @Fluent
 public final class PacketCaptureInner {
-    @JsonIgnore
-    private final ClientLogger logger = new ClientLogger(PacketCaptureInner.class);
+    @JsonIgnore private final ClientLogger logger = new ClientLogger(PacketCaptureInner.class);
 
     /*
-     * Parameters that define the create packet capture operation.
+     * Properties of the packet capture.
      */
     @JsonProperty(value = "properties", required = true)
     private PacketCaptureParameters innerProperties = new PacketCaptureParameters();
 
     /**
-     * Get the innerProperties property: Parameters that define the create
-     * packet capture operation.
-     * 
+     * Get the innerProperties property: Properties of the packet capture.
+     *
      * @return the innerProperties value.
      */
     private PacketCaptureParameters innerProperties() {
@@ -39,9 +33,8 @@ public final class PacketCaptureInner {
     }
 
     /**
-     * Get the target property: The ID of the targeted resource, only VM is
-     * currently supported.
-     * 
+     * Get the target property: The ID of the targeted resource, only VM is currently supported.
+     *
      * @return the target value.
      */
     public String target() {
@@ -49,9 +42,8 @@ public final class PacketCaptureInner {
     }
 
     /**
-     * Set the target property: The ID of the targeted resource, only VM is
-     * currently supported.
-     * 
+     * Set the target property: The ID of the targeted resource, only VM is currently supported.
+     *
      * @param target the target value to set.
      * @return the PacketCaptureInner object itself.
      */
@@ -64,23 +56,21 @@ public final class PacketCaptureInner {
     }
 
     /**
-     * Get the bytesToCapturePerPacket property: Number of bytes captured per
-     * packet, the remaining bytes are truncated.
-     * 
+     * Get the bytesToCapturePerPacket property: Number of bytes captured per packet, the remaining bytes are truncated.
+     *
      * @return the bytesToCapturePerPacket value.
      */
-    public Integer bytesToCapturePerPacket() {
+    public Long bytesToCapturePerPacket() {
         return this.innerProperties() == null ? null : this.innerProperties().bytesToCapturePerPacket();
     }
 
     /**
-     * Set the bytesToCapturePerPacket property: Number of bytes captured per
-     * packet, the remaining bytes are truncated.
-     * 
+     * Set the bytesToCapturePerPacket property: Number of bytes captured per packet, the remaining bytes are truncated.
+     *
      * @param bytesToCapturePerPacket the bytesToCapturePerPacket value to set.
      * @return the PacketCaptureInner object itself.
      */
-    public PacketCaptureInner withBytesToCapturePerPacket(Integer bytesToCapturePerPacket) {
+    public PacketCaptureInner withBytesToCapturePerPacket(Long bytesToCapturePerPacket) {
         if (this.innerProperties() == null) {
             this.innerProperties = new PacketCaptureParameters();
         }
@@ -89,23 +79,21 @@ public final class PacketCaptureInner {
     }
 
     /**
-     * Get the totalBytesPerSession property: Maximum size of the capture
-     * output.
-     * 
+     * Get the totalBytesPerSession property: Maximum size of the capture output.
+     *
      * @return the totalBytesPerSession value.
      */
-    public Integer totalBytesPerSession() {
+    public Long totalBytesPerSession() {
         return this.innerProperties() == null ? null : this.innerProperties().totalBytesPerSession();
     }
 
     /**
-     * Set the totalBytesPerSession property: Maximum size of the capture
-     * output.
-     * 
+     * Set the totalBytesPerSession property: Maximum size of the capture output.
+     *
      * @param totalBytesPerSession the totalBytesPerSession value to set.
      * @return the PacketCaptureInner object itself.
      */
-    public PacketCaptureInner withTotalBytesPerSession(Integer totalBytesPerSession) {
+    public PacketCaptureInner withTotalBytesPerSession(Long totalBytesPerSession) {
         if (this.innerProperties() == null) {
             this.innerProperties = new PacketCaptureParameters();
         }
@@ -114,9 +102,8 @@ public final class PacketCaptureInner {
     }
 
     /**
-     * Get the timeLimitInSeconds property: Maximum duration of the capture
-     * session in seconds.
-     * 
+     * Get the timeLimitInSeconds property: Maximum duration of the capture session in seconds.
+     *
      * @return the timeLimitInSeconds value.
      */
     public Integer timeLimitInSeconds() {
@@ -124,9 +111,8 @@ public final class PacketCaptureInner {
     }
 
     /**
-     * Set the timeLimitInSeconds property: Maximum duration of the capture
-     * session in seconds.
-     * 
+     * Set the timeLimitInSeconds property: Maximum duration of the capture session in seconds.
+     *
      * @param timeLimitInSeconds the timeLimitInSeconds value to set.
      * @return the PacketCaptureInner object itself.
      */
@@ -139,9 +125,8 @@ public final class PacketCaptureInner {
     }
 
     /**
-     * Get the storageLocation property: Describes the storage location for a
-     * packet capture session.
-     * 
+     * Get the storageLocation property: The storage location for a packet capture session.
+     *
      * @return the storageLocation value.
      */
     public PacketCaptureStorageLocation storageLocation() {
@@ -149,9 +134,8 @@ public final class PacketCaptureInner {
     }
 
     /**
-     * Set the storageLocation property: Describes the storage location for a
-     * packet capture session.
-     * 
+     * Set the storageLocation property: The storage location for a packet capture session.
+     *
      * @param storageLocation the storageLocation value to set.
      * @return the PacketCaptureInner object itself.
      */
@@ -164,8 +148,8 @@ public final class PacketCaptureInner {
     }
 
     /**
-     * Get the filters property: The filters property.
-     * 
+     * Get the filters property: A list of packet capture filters.
+     *
      * @return the filters value.
      */
     public List<PacketCaptureFilter> filters() {
@@ -173,8 +157,8 @@ public final class PacketCaptureInner {
     }
 
     /**
-     * Set the filters property: The filters property.
-     * 
+     * Set the filters property: A list of packet capture filters.
+     *
      * @param filters the filters value to set.
      * @return the PacketCaptureInner object itself.
      */
@@ -188,12 +172,15 @@ public final class PacketCaptureInner {
 
     /**
      * Validates the instance.
-     * 
+     *
      * @throws IllegalArgumentException thrown if the instance is not valid.
      */
     public void validate() {
         if (innerProperties() == null) {
-            throw logger.logExceptionAsError(new IllegalArgumentException("Missing required property innerProperties in model PacketCaptureInner"));
+            throw logger
+                .logExceptionAsError(
+                    new IllegalArgumentException(
+                        "Missing required property innerProperties in model PacketCaptureInner"));
         } else {
             innerProperties().validate();
         }
