+++ sdk/resourcemanager/azure-resourcemanager-network/src/main/java/com/azure/resourcemanager/network/models/ConnectionStateSnapshot.java
@@ -5,21 +5,16 @@
 package com.azure.resourcemanager.network.models;
 
 import com.azure.core.annotation.Fluent;
-import com.azure.core.annotation.Immutable;
 import com.azure.core.util.logging.ClientLogger;
-import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.time.OffsetDateTime;
 import java.util.List;
 
-/**
- * Connection state snapshot.
- */
+/** Connection state snapshot. */
 @Fluent
 public final class ConnectionStateSnapshot {
-    @JsonIgnore
-    private final ClientLogger logger = new ClientLogger(ConnectionStateSnapshot.class);
+    @JsonIgnore private final ClientLogger logger = new ClientLogger(ConnectionStateSnapshot.class);
 
     /*
      * The connection state.
@@ -49,31 +44,31 @@ public final class ConnectionStateSnapshot {
      * Average latency in ms.
      */
     @JsonProperty(value = "avgLatencyInMs")
-    private Integer avgLatencyInMs;
+    private Long avgLatencyInMs;
 
     /*
      * Minimum latency in ms.
      */
     @JsonProperty(value = "minLatencyInMs")
-    private Integer minLatencyInMs;
+    private Long minLatencyInMs;
 
     /*
      * Maximum latency in ms.
      */
     @JsonProperty(value = "maxLatencyInMs")
-    private Integer maxLatencyInMs;
+    private Long maxLatencyInMs;
 
     /*
      * The number of sent probes.
      */
     @JsonProperty(value = "probesSent")
-    private Integer probesSent;
+    private Long probesSent;
 
     /*
      * The number of failed probes.
      */
     @JsonProperty(value = "probesFailed")
-    private Integer probesFailed;
+    private Long probesFailed;
 
     /*
      * List of hops between the source and the destination.
@@ -83,7 +78,7 @@ public final class ConnectionStateSnapshot {
 
     /**
      * Get the connectionState property: The connection state.
-     * 
+     *
      * @return the connectionState value.
      */
     public ConnectionState connectionState() {
@@ -92,7 +87,7 @@ public final class ConnectionStateSnapshot {
 
     /**
      * Set the connectionState property: The connection state.
-     * 
+     *
      * @param connectionState the connectionState value to set.
      * @return the ConnectionStateSnapshot object itself.
      */
@@ -103,7 +98,7 @@ public final class ConnectionStateSnapshot {
 
     /**
      * Get the startTime property: The start time of the connection snapshot.
-     * 
+     *
      * @return the startTime value.
      */
     public OffsetDateTime startTime() {
@@ -112,7 +107,7 @@ public final class ConnectionStateSnapshot {
 
     /**
      * Set the startTime property: The start time of the connection snapshot.
-     * 
+     *
      * @param startTime the startTime value to set.
      * @return the ConnectionStateSnapshot object itself.
      */
@@ -123,7 +118,7 @@ public final class ConnectionStateSnapshot {
 
     /**
      * Get the endTime property: The end time of the connection snapshot.
-     * 
+     *
      * @return the endTime value.
      */
     public OffsetDateTime endTime() {
@@ -132,7 +127,7 @@ public final class ConnectionStateSnapshot {
 
     /**
      * Set the endTime property: The end time of the connection snapshot.
-     * 
+     *
      * @param endTime the endTime value to set.
      * @return the ConnectionStateSnapshot object itself.
      */
@@ -142,9 +137,8 @@ public final class ConnectionStateSnapshot {
     }
 
     /**
-     * Get the evaluationState property: Connectivity analysis evaluation
-     * state.
-     * 
+     * Get the evaluationState property: Connectivity analysis evaluation state.
+     *
      * @return the evaluationState value.
      */
     public EvaluationState evaluationState() {
@@ -152,9 +146,8 @@ public final class ConnectionStateSnapshot {
     }
 
     /**
-     * Set the evaluationState property: Connectivity analysis evaluation
-     * state.
-     * 
+     * Set the evaluationState property: Connectivity analysis evaluation state.
+     *
      * @param evaluationState the evaluationState value to set.
      * @return the ConnectionStateSnapshot object itself.
      */
@@ -165,108 +158,107 @@ public final class ConnectionStateSnapshot {
 
     /**
      * Get the avgLatencyInMs property: Average latency in ms.
-     * 
+     *
      * @return the avgLatencyInMs value.
      */
-    public Integer avgLatencyInMs() {
+    public Long avgLatencyInMs() {
         return this.avgLatencyInMs;
     }
 
     /**
      * Set the avgLatencyInMs property: Average latency in ms.
-     * 
+     *
      * @param avgLatencyInMs the avgLatencyInMs value to set.
      * @return the ConnectionStateSnapshot object itself.
      */
-    public ConnectionStateSnapshot withAvgLatencyInMs(Integer avgLatencyInMs) {
+    public ConnectionStateSnapshot withAvgLatencyInMs(Long avgLatencyInMs) {
         this.avgLatencyInMs = avgLatencyInMs;
         return this;
     }
 
     /**
      * Get the minLatencyInMs property: Minimum latency in ms.
-     * 
+     *
      * @return the minLatencyInMs value.
      */
-    public Integer minLatencyInMs() {
+    public Long minLatencyInMs() {
         return this.minLatencyInMs;
     }
 
     /**
      * Set the minLatencyInMs property: Minimum latency in ms.
-     * 
+     *
      * @param minLatencyInMs the minLatencyInMs value to set.
      * @return the ConnectionStateSnapshot object itself.
      */
-    public ConnectionStateSnapshot withMinLatencyInMs(Integer minLatencyInMs) {
+    public ConnectionStateSnapshot withMinLatencyInMs(Long minLatencyInMs) {
         this.minLatencyInMs = minLatencyInMs;
         return this;
     }
 
     /**
      * Get the maxLatencyInMs property: Maximum latency in ms.
-     * 
+     *
      * @return the maxLatencyInMs value.
      */
-    public Integer maxLatencyInMs() {
+    public Long maxLatencyInMs() {
         return this.maxLatencyInMs;
     }
 
     /**
      * Set the maxLatencyInMs property: Maximum latency in ms.
-     * 
+     *
      * @param maxLatencyInMs the maxLatencyInMs value to set.
      * @return the ConnectionStateSnapshot object itself.
      */
-    public ConnectionStateSnapshot withMaxLatencyInMs(Integer maxLatencyInMs) {
+    public ConnectionStateSnapshot withMaxLatencyInMs(Long maxLatencyInMs) {
         this.maxLatencyInMs = maxLatencyInMs;
         return this;
     }
 
     /**
      * Get the probesSent property: The number of sent probes.
-     * 
+     *
      * @return the probesSent value.
      */
-    public Integer probesSent() {
+    public Long probesSent() {
         return this.probesSent;
     }
 
     /**
      * Set the probesSent property: The number of sent probes.
-     * 
+     *
      * @param probesSent the probesSent value to set.
      * @return the ConnectionStateSnapshot object itself.
      */
-    public ConnectionStateSnapshot withProbesSent(Integer probesSent) {
+    public ConnectionStateSnapshot withProbesSent(Long probesSent) {
         this.probesSent = probesSent;
         return this;
     }
 
     /**
      * Get the probesFailed property: The number of failed probes.
-     * 
+     *
      * @return the probesFailed value.
      */
-    public Integer probesFailed() {
+    public Long probesFailed() {
         return this.probesFailed;
     }
 
     /**
      * Set the probesFailed property: The number of failed probes.
-     * 
+     *
      * @param probesFailed the probesFailed value to set.
      * @return the ConnectionStateSnapshot object itself.
      */
-    public ConnectionStateSnapshot withProbesFailed(Integer probesFailed) {
+    public ConnectionStateSnapshot withProbesFailed(Long probesFailed) {
         this.probesFailed = probesFailed;
         return this;
     }
 
     /**
-     * Get the hops property: List of hops between the source and the
-     * destination.
-     * 
+     * Get the hops property: List of hops between the source and the destination.
+     *
      * @return the hops value.
      */
     public List<ConnectivityHop> hops() {
@@ -275,7 +267,7 @@ public final class ConnectionStateSnapshot {
 
     /**
      * Validates the instance.
-     * 
+     *
      * @throws IllegalArgumentException thrown if the instance is not valid.
      */
     public void validate() {
